+++ R/get-generics.R
@@ -13,6 +13,21 @@ get_region <- function(x, ...) {
   UseMethod("get_region")
 }
 
+#' Get the all included regions' name
+#'
+#' This is a generic function aimed to access to the all the regions' name
+#' into an object
+#'
+#' @param x an object used to select a method.
+#' @param ... further arguments passed to or from other methods.
+#'
+#' @return The regions' name for the object \code{x}.
+#'
+#' @export
+get_all_region <- function(x, ...) {
+  UseMethod("get_all_region")
+}
+
 #' Get the probability of acceptance
 #'
 #' This is a generic function aimed to access (or compute) to the
@@ -97,6 +112,21 @@ get_regions <- function(x, ...) {
   UseMethod("get_regions")
 }
 
+#' Get the macroregions details
+#'
+#' This is a generic function aimed to access to the \code{macroregions}
+#' details of an object
+#'
+#' @param x an object used to select a method.
+#' @param ... further arguments passed to or from other methods.
+#'
+#' @return The \code{macroregions} details for the object \code{x}.
+#'
+#' @export
+get_macroregions <- function(x, ...) {
+  UseMethod("get_macroregions")
+}
+
 #' Get the initial strip details
 #'
 #' This is a generic function aimed to access to the \code{initial_strip}
@@ -126,16 +156,44 @@ get_current_strip <- function(x, ...) {
   UseMethod("get_current_strip")
 }
 
-#' Get the current time details
+#' Get the initial time details
 #'
-#' This is a generic function aimed to access to the \code{time_period}
+#' This is a generic function aimed to access to the \code{initial_time}
 #' details of an object.
 #'
 #' @param x an object used to select a method.
 #' @param ... further arguments passed to or from other methods.
 #'
-#' @return The \code{time_period} details for the object \code{x}.
+#' @return The \code{initial_time} details for the object \code{x}.
+#' @export
+get_initial_time <- function(x, ...) {
+  UseMethod("get_initial_time")
+}
+
+#' Get the current time details
+#'
+#' This is a generic function aimed to access to the \code{current_time}
+#' details of an object.
+#'
+#' @param x an object used to select a method.
+#' @param ... further arguments passed to or from other methods.
+#'
+#' @return The \code{current_time} details for the object \code{x}.
 #' @export
 get_time <- function(x, ...) {
   UseMethod("get_time")
+}
+
+#' Get the final time details
+#'
+#' This is a generic function aimed to access to the \code{final_time}
+#' details of an object.
+#'
+#' @param x an object used to select a method.
+#' @param ... further arguments passed to or from other methods.
+#'
+#' @return The \code{final_time} details for the object \code{x}.
+#' @export
+get_final_time <- function(x, ...) {
+  UseMethod("get_final_time")
 }